+++ src/projectgoldstarscalculators/Main.java
@@ -4,7 +4,7 @@ import java.awt.Font;
 import java.io.File;
 public class Main
 {
-    public static final String NAME = "Project GoldStars Calculator S 3.1 Development Build 1";
+    public static final String NAME = "Project GoldStars Calculator S 3.1 Development Build 2";
     public static final String VERSION = "Version 3.1";
     public static final File FOLDER = new File(System.getProperty("user.home"), "Project GoldStars Calculator S");
     public static int height;   //Height of the screen (in pixels)